requirements.txt
alembic==1.9.1
attrs==22.2.0
Babel==2.11.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.1
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.1
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.1
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
